Move src/index.ts to hardhat-scripts/constants/protocolConstants.ts
